Created the ecgstep mapclient plugin.
This plugin is used for registering Blackfynn data to a given fitted mesh of ECG points found on a heart.

Input is from a previous step containing the coordinates of the nodes defining the mesh.

This repository is currently heavily in development. I will try add more documentation as I go